Rename package to `httpchunked`
